mytodolist/requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
uritemplate==3.0.1
